+++ src/containers/SearchPage/searchSelectors.js
@@ -9,7 +9,7 @@
 import { createSelector } from 'reselect';
 import { constants } from '@ndla/ui';
 import { convertFieldWithFallback } from '../../util/convertFieldWithFallback';
-import getContentTypeFromResourceTypes from '../../util/getContentTypeFromResourceTypes';
+import getContentTypeFromResourceTypes from '../../util/getContentType';
 import config from '../../config';
 
 const getContentType = resource => {
