+++ src/util/getContentType.jsx
